Add identity check back

+++ update.rb
@@ -11,18 +11,18 @@ Dir.chdir pwd
 
 docs.each do |doc|
   # puts "Checking #{doc}"
-  # unless (File.exist? "#{doc}.orig") && (FileUtils.identical? "#{emacs}/doc/#{doc}", "#{doc}.orig")
-  FileUtils.copy "#{emacs}/doc/#{doc}", "#{doc}.orig"
-  puts "Updated #{doc}"
-  file = Nokogiri::HTML(File.open("#{doc}.orig"))
-  nodes = file.search(".node")
-  nodes.each do |node|
-    link = node.css("a[name]")
-    node.children = link
+  unless (File.exist? "#{doc}.orig") && (FileUtils.identical? "#{emacs}/doc/#{doc}", "#{doc}.orig")
+    FileUtils.copy "#{emacs}/doc/#{doc}", "#{doc}.orig"
+    puts "Updated #{doc}"
+    file = Nokogiri::HTML(File.open("#{doc}.orig"))
+    nodes = file.search(".node")
+    nodes.each do |node|
+      link = node.css("a[name]")
+      node.children = link
+    end
+    File.open(doc, "w") { |f|
+      f.puts file
+    }
+    puts "Generating #{doc}"
   end
-  File.open(doc, "w") { |f|
-    f.puts file
-  }
-  puts "Generating #{doc}"
-  # end
 end
